refactor: implement code review feedback
