+++ src/app/App.tsx
@@ -1,9 +1,10 @@
-import { Provider } from './Provider';
+import { RouterProvider } from 'react-router-dom';
+import router from './Router';
 
 function App() {
   return (
     <>
-      <Provider />
+      <RouterProvider router={router} />
     </>
   );
 }
