--- src/app/Layout/Footer.tsx
@@ -1,3 +1,0 @@
-export const Footer = () => {
-  return <footer>footer</footer>;
-};
