+++ src/app/Layout/Layout.tsx
@@ -1,15 +1,9 @@
 import { Outlet } from 'react-router-dom';
-import { Navbar } from './Navbar';
-import { Sidebar } from './Sidebar';
-import { Footer } from './Footer';
 
 export const Layout = () => {
   return (
     <div>
-      <Navbar />
-      <Sidebar />
       <Outlet />
-      <Footer />
     </div>
   );
 };
