--- src/app/Layout/Navbar.tsx
@@ -1,3 +1,0 @@
-export const Navbar = () => {
-  return <nav>navbar</nav>;
-};
