--- src/app/Layout/Sidebar.tsx
@@ -1,3 +1,0 @@
-export const Sidebar = () => {
-  return <div>sidebar</div>;
-};
