--- src/app/Provider.tsx
@@ -1,6 +1,0 @@
-import { RouterProvider } from 'react-router-dom';
-import router from './Router';
-
-export const Provider = () => {
-  return <RouterProvider router={router} />;
-};
