--- src/features/TransactionsTable/components/TransactionTable.tsx
@@ -1,3 +1,0 @@
-export const TransactionsTable = () => {
-  return <>Transactions</>;
-};
