+++ src/features/transactionTable/pages/TransactionTable.tsx
@@ -1,0 +1,3 @@
+export const TransactionTable = () => {
+  return <>Transactions</>;
+};
